NODE-665: binops to become functions
